requirements.txt
flake8==5.0.4
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
